Began work on new feature
Cloning into a stage instead of directly into a home makes this script
significantly more useable (as it is rare for even a brand new user to
have an empty home) as well as being a good practice.

The changes in this first commit do most of the work for setting this up
(it's not that complicated), but there's still a problem where symlinks
created by sub-setup scripts (the ones that live in dotdir) point to the
stage instead of the real home, which becomes a problem after all the
files are moved into the real home and the stage is destroyed.